+++ views/teams/form.php
@@ -35,7 +35,7 @@ $v = function ($key, $default = '') use ($team) {
     <div class="col-md-6">
       <label class="form-label">Telegram Chat ID ομάδας</label>
       <input type="text" name="telegram_chat_id" class="form-control" value="<?= e($v('telegram_chat_id')) ?>" placeholder="π.χ. -1001234567890">
-      <div class="form-text">Προαιρετικό group/channel της ομάδας για Telegram ειδοποιήσεις.</div>
+      <div class="form-text">Προαιρετικό. Συμπληρώστε μόνο αν αυτή η ομάδα έχει δικό της ξεχωριστό Telegram group. Αλλιώς χρησιμοποιείται το κοινό Chat ID ομάδων από τις Ρυθμίσεις.</div>
     </div>
     <div class="col-md-6">
       <label class="form-label">Διεύθυνση</label>